refactor: extract create transaction use case
