+++ src/test/java/com/financebot/transaction/service/TransactionServiceTest.java
@@ -10,11 +10,11 @@ import com.financebot.transaction.domain.installment.InstallmentPlan;
 import com.financebot.transaction.domain.installment.InstallmentPlanFactory;
 import com.financebot.transaction.domain.installment.InstallmentPlanItem;
 import com.financebot.transaction.dto.TransactionFilter;
-import com.financebot.transaction.dto.request.CreateInstallmentTransactionRequest;
-import com.financebot.transaction.dto.request.CreateTransactionRequest;
-import com.financebot.transaction.dto.request.UpdateTransactionRequest;
-import com.financebot.transaction.dto.response.InstallmentTransactionResponse;
-import com.financebot.transaction.dto.response.TransactionResponse;
+import com.financebot.transaction.application.dto.request.CreateInstallmentTransactionRequest;
+import com.financebot.transaction.application.dto.request.CreateTransactionRequest;
+import com.financebot.transaction.application.dto.request.UpdateTransactionRequest;
+import com.financebot.transaction.application.dto.response.InstallmentTransactionResponse;
+import com.financebot.transaction.application.dto.response.TransactionResponse;
 import com.financebot.transaction.mapper.TransactionMapper;
 import com.financebot.transaction.repository.TransactionRepository;
 import com.financebot.transaction.validation.TransactionCategoryValidator;
